refactor: remove view folder in notice feature
