+++ src/app/notice/page.tsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { type Notice } from '@/entities/notice/model/types';
-import { CreateNoticeCard } from '@/features/notice/view/ui/CreateNoticeCard';
-import NoticeCard from '@/features/notice/view/ui/NoticeCard';
+import { CreateNoticeCard } from '@/features/notice/ui/CreateNoticeCard';
+import NoticeCard from '@/features/notice/ui/NoticeCard';
 import { createNoticeForm } from '@/shared/api/createNoticeForm';
 import { fetchNotices } from '@/shared/api/fetchNotices';
 import { useEffect, useState } from 'react';
 import { deleteNotice } from '@/shared/api/deleteNotice';
 import { mockNotices } from '@/shared/mock/notices';
-import { useCreateNotice } from '@/features/notice/view/lib/useCreateNotice';
+import { useCreateNotice } from '@/features/notice/lib/useCreateNotice';
 export default function Notice() {
   const [notices, setNotices] = useState<Notice[]>([]);
