+++ src/features/notice/ui/NoticeCard.tsx
