+++ src/features/notice/ui/NoticeDetailCard.tsx
